Major rearrangement of src/*.py files to fit a python package
- All files moved to under src/flipjump/ which will be the package
- organized *.py files into different folders under src/flipjump/
- updated all imports to be absolute imports
- split fjm.py into fjm/{fjm_consts, fjm_reader, fjm_writer}.py
- split defs.py into utils/{classes, constants, functions}.py
- fix general python-warnings
- optimize imports
