+++ .gitignore
@@ -16,9 +16,6 @@ __pycache__/
 *.py[cod]
 *$py.class
 
-# C extensions
-*.so
-
 # Distribution / packaging
 .Python
 build/
@@ -156,7 +153,7 @@ dmypy.json
 *.gch
 *.pch
 
-# Compiled Dynamic libraries
+# Compiled Dynamic libraries (also "C extensions")
 *.so
 *.dylib
 *.dll
