--- src/__init__.py
@@ -1,5 +1,0 @@
-from os.path import abspath, dirname
-import sys
-
-
-sys.path.append(abspath(dirname(__file__)))
