+++ src/flipjump/fjm/fjm_consts.py
@@ -1,0 +1,62 @@
+import lzma
+from typing import List, Dict
+
+
+"""
+struct {
+    u16 fj_magic;   // 'F' + 'J'<<8  (0x4a46)
+    u16 word_size;  // in bits
+    u64 version;
+    u64 segment_num;
+    { // for versions > 0
+        u64 flags;
+        u32 reserved;   // 0
+    }
+    struct segment {
+        u64 segment_start;  // in memory words (w-bits)
+        u64 segment_length; // in memory words (w-bits)
+        u64 data_start;     // in the outer-struct.data words (w-bits)
+        u64 data_length;    // in the outer-struct.data words (w-bits)
+    } *segments;        // segments[segment_num]
+    u8* data;       // the data (might be compressed in some versions)
+} fjm_file;     // Flip-Jump Memory file
+"""
+
+
+FJ_MAGIC = ord('F') + (ord('J') << 8)
+
+
+_reserved_dict_threshold = 1000
+
+_header_base_format = '<HHQQ'
+_header_base_size = 2 + 2 + 8 + 8
+
+_header_extension_format = '<QL'
+_header_extension_size = 8 + 4
+
+_segment_format = '<QQQQ'
+_segment_size = 8 + 8 + 8 + 8
+
+
+BaseVersion = 0
+NormalVersion = 1
+RelativeJumpVersion = 2     # compress-friendly
+CompressedVersion = 3       # version 2 but data is lzma2-compressed
+SUPPORTED_VERSIONS = {
+    BaseVersion: 'Base',
+    NormalVersion: 'Normal',
+    RelativeJumpVersion: 'RelativeJump',
+    CompressedVersion: 'Compressed',
+}
+
+
+_LZMA_FORMAT = lzma.FORMAT_RAW
+_LZMA_DECOMPRESSION_FILTERS: List[Dict[str, int]] = [{"id": lzma.FILTER_LZMA2}]
+
+
+def _lzma_compression_filters(dw: int, preset: int) -> List[Dict[str, int]]:
+    return [{"id": lzma.FILTER_LZMA2, "preset": preset, "nice_len": dw}]
+
+
+def _new_garbage_val() -> int:
+    return 0        # The value read when reading a word outside any segment.
