+++ src/flipjump/inner_classes/exceptions.py
