+++ src/flipjump/io_devices/BrokenIO.py
@@ -1,5 +1,5 @@
-from .IODevice import IODevice
-from .io_exceptions import BrokenIOUsed
+from flipjump.io_devices.IODevice import IODevice
+from flipjump.io_devices.io_exceptions import BrokenIOUsed
 
 
 class BrokenIO(IODevice):
